QoL changes, more consitent naming, added some comments to functions that probably wont change their signatures, split statsd-parsers into two files parsers/network-listener due to separation of concerns, split up a lot of definitions from what was statsd-parsers.h to multiple files as all threads args and their constructors were defined in it (and more)
