+++ src/aggregator-counter.c
@@ -1,7 +1,7 @@
 #include <string.h>
 
 #include "config-reader.h"
-#include "statsd-parsers.h"
+#include "network-listener.h"
 #include "aggregators.h"
 #include "aggregator-counter.h"
 #include "errno.h"
