+++ src/aggregator-counter.h
@@ -2,8 +2,9 @@
 #define AGGREGATOR_COUNTER_
 
 #include <stdio.h>
+
 #include "config-reader.h"
-#include "statsd-parsers.h"
+#include "network-listener.h"
 #include "aggregators.h"
 
 /**
