+++ src/aggregator-duration.c
@@ -2,7 +2,7 @@
 #include <string.h>
 
 #include "config-reader.h"
-#include "statsd-parsers.h"
+#include "network-listener.h"
 #include "aggregators.h"
 #include "aggregator-duration.h"
 #include "aggregator-duration-exact.h"
