+++ src/aggregator-duration.h
@@ -2,7 +2,7 @@
 #define AGGREGATOR_DURATION_
 
 #include "config-reader.h"
-#include "statsd-parsers.h"
+#include "network-listener.h"
 #include "aggregators.h"
 #include "aggregator-duration.h"
 #include "aggregator-duration-exact.h"
