+++ src/aggregator-gauge.c
@@ -2,7 +2,7 @@
 #include <string.h>
 
 #include "config-reader.h"
-#include "statsd-parsers.h"
+#include "network-listener.h"
 #include "aggregators.h"
 #include "aggregator-gauge.h"
 #include "errno.h"
