+++ src/config-reader.h
@@ -33,12 +33,31 @@ typedef struct agent_config {
     char* tcp_address;
 } agent_config;
 
+/**
+ * Read agent config from either file or command line arguments
+ * @arg src_flag - Specifies config source, 0 = READ_FROM_FILE, 1 = READ_FROM_CMD
+ * @arg config_path - Path to config file
+ * @return Program configuration
+ */
 agent_config* read_agent_config(int src_flag, char* config_path, int argc, char **argv);
 
+/**
+ * Reads program config from given path
+ * @arg agent_config - Placeholder config to write what was read to
+ * @arg path - Path to read file from
+ */
 void read_agent_config_file(agent_config** dest, char* path);
 
+/**
+ * Reads program config from command line arguments
+ * @arg agent_config - Placeholder config to write what was read to
+ */
 void read_agent_config_cmd(agent_config** dest, int argc, char **argv);
 
+/**
+ * Print out agent config to STDOUT
+ * @arg config - Config to print out
+ */
 void print_agent_config(agent_config* config);
 
 #endif
